requirements.txt
Flask==2.2.2
whisper==1.0.0  # Ensure you have the correct version for your Whisper implementation
Werkzeug==2.2.2  # Adjust based on compatibility
langchain==0.2.17 
langchain-core==0.2.43 
langsmith==0.1.147
langchain-community==0.2.19
flask-cors==3.0.10  # For CORS support
soundfile==0.10.2  # For reading and writing audio files
numpy==1.24.4  # Required for numerical operations
pydub==0.25.1
pydantic==2.10.6
flask-socketio==5.5.1
opencc-python-reimplemented==0.1.6  # Specify version for compatibility

